Changing Units::Prefix from double to int

+++ src/api/libcellml/units.h
@@ -151,10 +151,10 @@ public:
      * @param multiplier The multiplier.
      * @param id The optional @c std::string id to set for this unit.
      */
-    //void addUnit(const std::string &reference, int prefix, double exponent,
-    //             double multiplier=1.0, const std::string &id="");
-    void addUnit(const std::string &reference, double prefix, double exponent,
+    void addUnit(const std::string &reference, int prefix, double exponent,
                  double multiplier=1.0, const std::string &id="");
+    //void addUnit(const std::string &reference, double prefix, double exponent,
+    //             double multiplier=1.0, const std::string &id="");
 
     /**
      * @brief Add a unit to this Units.
@@ -230,10 +230,10 @@ public:
      * @param multiplier The multiplier.
      * @param id The optional @c std::string id to set for this unit.
      */
-    //void addUnit(StandardUnit standardRef, int prefix, double exponent,
-    //             double multiplier=1.0, const std::string &id="");
-    void addUnit(StandardUnit standardRef, double prefix, double exponent,
+    void addUnit(StandardUnit standardRef, int prefix, double exponent,
                  double multiplier=1.0, const std::string &id="");
+    //void addUnit(StandardUnit standardRef, double prefix, double exponent,
+    //             double multiplier=1.0, const std::string &id="");
 
     /**
      * @brief Add a unit to this Units.
